+++ src/main/java/net/geforcemods/securitycraft/gui/components/ColorableScrollPanel.java
@@ -14,7 +14,22 @@ import net.minecraft.client.renderer.vertex.DefaultVertexFormats;
 import net.minecraftforge.fml.client.config.GuiUtils;
 
 // from Forge, changed scrollbar/-list color and removed things unnecessary for this usecase
-public abstract class PatronScrollList {
+public abstract class ColorableScrollPanel {
+	public static class Color {
+		public final int r, g, b, a;
+
+		public Color(int r, int g, int b, int a) {
+			this.r = r;
+			this.g = g;
+			this.b = b;
+			this.a = a;
+		}
+
+		public int combinedRGBA() {
+			return ((r & 0xFF) << 24) | ((g & 0xFF) << 16) | ((b & 0xFF) << 8) | (a & 0xFF);
+		}
+	}
+
 	private final Minecraft client;
 	protected final int listWidth;
 	protected final int listHeight;
@@ -34,8 +49,17 @@ public abstract class PatronScrollList {
 	private int headerHeight;
 	protected boolean captureMouse = true;
 	public boolean isHovering;
+	private Color backgroundTo;
+	private Color backgroundFrom;
+	private Color scrollbarBackground;
+	private Color scrollbarBorder;
+	private Color scrollbar;
 
-	public PatronScrollList(Minecraft client, int width, int height, int top, int bottom, int left, int entryHeight, int screenWidth, int screenHeight) {
+	public ColorableScrollPanel(Minecraft client, int width, int height, int top, int bottom, int left, int entryHeight, int screenWidth, int screenHeight, Color backgroundFrom, Color backgroundTo) {
+		this(client, width, height, top, bottom, left, entryHeight, screenWidth, screenHeight, backgroundFrom, backgroundTo, new Color(0x00, 0x00, 0x00, 0xFF), new Color(0x80, 0x80, 0x80, 0xFF), new Color(0xC0, 0xC0, 0xC0, 0xFF));
+	}
+
+	public ColorableScrollPanel(Minecraft client, int width, int height, int top, int bottom, int left, int entryHeight, int screenWidth, int screenHeight, Color backgroundFrom, Color backgroundTo, Color scrollbarBackground, Color scrollbarBorder, Color scrollbar) {
 		this.client = client;
 		listWidth = width;
 		listHeight = height;
@@ -46,6 +70,11 @@ public abstract class PatronScrollList {
 		right = width + left;
 		this.screenWidth = screenWidth;
 		this.screenHeight = screenHeight;
+		this.backgroundFrom = backgroundFrom;
+		this.backgroundTo = backgroundTo;
+		this.scrollbarBackground = scrollbarBackground;
+		this.scrollbarBorder = scrollbarBorder;
+		this.scrollbar = scrollbar;
 	}
 
 	public abstract int getSize();
@@ -102,7 +131,7 @@ public abstract class PatronScrollList {
 		applyScrollLimits();
 		GL11.glEnable(GL11.GL_SCISSOR_TEST);
 		GL11.glScissor((int) (left * scaleW), (int) (client.displayHeight - (bottom * scaleH)), (int) (listWidth * scaleW), (int) (viewHeight * scaleH));
-		GuiUtils.drawGradientRect(0, left, top, right, bottom, 0xC0BFBBB2, 0xD0BFBBB2); //list background
+		GuiUtils.drawGradientRect(0, left, top, right, bottom, backgroundFrom.combinedRGBA(), backgroundTo.combinedRGBA()); //list background
 		drawPanel(right, top + border - (int) scrollDistance, tess, mouseX, mouseY);
 		GlStateManager.disableDepth();
 
@@ -110,31 +139,30 @@ public abstract class PatronScrollList {
 			int height = getBarHeight(viewHeight, border);
 			int barTop = (int) scrollDistance * (viewHeight - height) / extraHeight + top;
 
-			if (barTop < top) {
+			if (barTop < top)
 				barTop = top;
-			}
 
 			GlStateManager.disableTexture2D();
 			//scrollbar background
 			worldr.begin(GL11.GL_QUADS, DefaultVertexFormats.POSITION_TEX_COLOR);
-			worldr.pos(scrollBarLeft, bottom, 0.0D).tex(0.0D, 1.0D).color(0x8E, 0x82, 0x70, 0xFF).endVertex();
-			worldr.pos(scrollBarRight, bottom, 0.0D).tex(1.0D, 1.0D).color(0x8E, 0x82, 0x70, 0xFF).endVertex();
-			worldr.pos(scrollBarRight, top, 0.0D).tex(1.0D, 0.0D).color(0x8E, 0x82, 0x70, 0xFF).endVertex();
-			worldr.pos(scrollBarLeft, top, 0.0D).tex(0.0D, 0.0D).color(0x8E, 0x82, 0x70, 0xFF).endVertex();
+			worldr.pos(scrollBarLeft, bottom, 0.0D).tex(0.0D, 1.0D).color(scrollbarBackground.r, scrollbarBackground.g, scrollbarBackground.b, scrollbarBackground.a).endVertex();
+			worldr.pos(scrollBarRight, bottom, 0.0D).tex(1.0D, 1.0D).color(scrollbarBackground.r, scrollbarBackground.g, scrollbarBackground.b, scrollbarBackground.a).endVertex();
+			worldr.pos(scrollBarRight, top, 0.0D).tex(1.0D, 0.0D).color(scrollbarBackground.r, scrollbarBackground.g, scrollbarBackground.b, scrollbarBackground.a).endVertex();
+			worldr.pos(scrollBarLeft, top, 0.0D).tex(0.0D, 0.0D).color(scrollbarBackground.r, scrollbarBackground.g, scrollbarBackground.b, scrollbarBackground.a).endVertex();
 			tess.draw();
 			//scrollbar border
 			worldr.begin(GL11.GL_QUADS, DefaultVertexFormats.POSITION_TEX_COLOR);
-			worldr.pos(scrollBarLeft, barTop + height, 0.0D).tex(0.0D, 1.0D).color(0x80, 0x70, 0x55, 0xFF).endVertex();
-			worldr.pos(scrollBarRight, barTop + height, 0.0D).tex(1.0D, 1.0D).color(0x80, 0x70, 0x55, 0xFF).endVertex();
-			worldr.pos(scrollBarRight, barTop, 0.0D).tex(1.0D, 0.0D).color(0x80, 0x70, 0x55, 0xFF).endVertex();
-			worldr.pos(scrollBarLeft, barTop, 0.0D).tex(0.0D, 0.0D).color(0x80, 0x70, 0x55, 0xFF).endVertex();
+			worldr.pos(scrollBarLeft, barTop + height, 0.0D).tex(0.0D, 1.0D).color(scrollbarBorder.r, scrollbarBorder.g, scrollbarBorder.b, scrollbarBorder.a).endVertex();
+			worldr.pos(scrollBarRight, barTop + height, 0.0D).tex(1.0D, 1.0D).color(scrollbarBorder.r, scrollbarBorder.g, scrollbarBorder.b, scrollbarBorder.a).endVertex();
+			worldr.pos(scrollBarRight, barTop, 0.0D).tex(1.0D, 0.0D).color(scrollbarBorder.r, scrollbarBorder.g, scrollbarBorder.b, scrollbarBorder.a).endVertex();
+			worldr.pos(scrollBarLeft, barTop, 0.0D).tex(0.0D, 0.0D).color(scrollbarBorder.r, scrollbarBorder.g, scrollbarBorder.b, scrollbarBorder.a).endVertex();
 			tess.draw();
 			//scrollbar
 			worldr.begin(GL11.GL_QUADS, DefaultVertexFormats.POSITION_TEX_COLOR);
-			worldr.pos(scrollBarLeft, barTop + height - 1, 0.0D).tex(0.0D, 1.0D).color(0xD1, 0xBF, 0xA1, 0xFF).endVertex();
-			worldr.pos(scrollBarRight - 1, barTop + height - 1, 0.0D).tex(1.0D, 1.0D).color(0xD1, 0xBF, 0xA1, 0xFF).endVertex();
-			worldr.pos(scrollBarRight - 1, barTop, 0.0D).tex(1.0D, 0.0D).color(0xD1, 0xBF, 0xA1, 0xFF).endVertex();
-			worldr.pos(scrollBarLeft, barTop, 0.0D).tex(0.0D, 0.0D).color(0xD1, 0xBF, 0xA1, 0xFF).endVertex();
+			worldr.pos(scrollBarLeft, barTop + height - 1, 0.0D).tex(0.0D, 1.0D).color(scrollbar.r, scrollbar.g, scrollbar.b, scrollbar.a).endVertex();
+			worldr.pos(scrollBarRight - 1, barTop + height - 1, 0.0D).tex(1.0D, 1.0D).color(scrollbar.r, scrollbar.g, scrollbar.b, scrollbar.a).endVertex();
+			worldr.pos(scrollBarRight - 1, barTop, 0.0D).tex(1.0D, 0.0D).color(scrollbar.r, scrollbar.g, scrollbar.b, scrollbar.a).endVertex();
+			worldr.pos(scrollBarLeft, barTop, 0.0D).tex(0.0D, 0.0D).color(scrollbar.r, scrollbar.g, scrollbar.b, scrollbar.a).endVertex();
 			tess.draw();
 		}