+++ src/main/java/net/geforcemods/securitycraft/tileentity/TileEntityProjector.java
@@ -180,9 +180,10 @@ public class TileEntityProjector extends TileEntityDisguisable implements IInven
 		if (count >= 1) {
 			projectedBlock = ItemStack.EMPTY;
 			resetSavedState();
+			return stack;
 		}
 
-		return stack;
+		return ItemStack.EMPTY;
 	}
 
 	@Override